Revise types used for nodes in TreeListModel.
Rather than a single TreeNode struct with some Option fields, define
separate RootNode and ItemNode structs, which both implement a Node
trait. This avoids needing to deal with an Option in contexts where we
know we are dealing with an item node and not the root.

Also define type aliases for the common Rc<RefCell<...>> wrappers.