Revise the ItemSource API, adding completion status.
When obtaining the count of children for an item, a CompletionStatus is
now also returned, indicating whether the count is final or may increase
in future.

While revising the API, we also modify some method signatures to take
Option<&Item> rather than &Option<Item>, which avoids copying existing
Item values to create an Option.