Keep track of incomplete items that may gain children.
Incomplete items are tracked from when item nodes for them are created.
Those nodes may persist as long as they are used by either the
TreeListModel or the ListView. We store weak references to those which
are known to be incomplete, so that we can check for new children on
update.

The references are stored hierarchically, as a BTreeMap of incomplete
children at each tree node, indexed by their item index relative to the
parent.

Note that this commit does not include code to remove nodes from their
parent's map after they are completed or no longer referenced. This is
done in the next commit.