Make the update method not require a &mut reference.
Although logically the update modifies the tree model, a &mut reference
is not required, because the updated data is stored in the RefCells that
hold each tree node.

Using a non-mutable borrow of the RefCell that holds the TreeListModel
will allow us to move the items_changed() call inside the update method.