Stop periodic updates when capture completes.
We add a finish() method to flag a Capture as complete, and call this
after decoding the last packet.

Once the capture is complete, the ItemSource::item_children() methods
will return a completion status of Complete for the root item.

When the TreeListModel update method sees that the capture is complete,
and it has handled all updates required to the tree model, it returns
false to indicate that there will be no further updates to that model.

When there are no further updates expected from either tree model, the
UI will stop scheduling the update routine.