When an item's summary must change, update its widget directly.
This change is necessary because previously it was impossible to expand
a row while its summary kept changing.

This happened because when we called items_changed(), the widget
representing the row would be removed and replaced by a different one.
Press and release events would land on different widgets, and fail to
register a click event on any indivdual widget.

Instead, we now register widgets with the node they are bound to, and
when the summary needs to be updated, we just replace the text rather
than calling into items_changed(). This also speeds up the update, since
the item no longer has to be re-fetched and bound to a new widget.

A HashSet is used to allow multiple widgets to be registered to the same
node, because it turns out that the GTK ListView implementation may have
multiple widgets bound to the same node at one time, and the most recent
one is not always the visible one.

The HashSet is wrapped in a RefCell because we cannot provide exclusive
access to the whole node at the point where a bind/unbind occurs.