Separate main.rs into lib.rs, ui.rs, and test_replay.rs.
All the statements declaring submodules are now in lib.rs, which
defines a packetry crate, used by main.rs. The UI code is moved to
ui.rs, which is exposed to main.rs as part of the packetry crate.

The UI replay test is now a separate standalone program, test_replay.rs.
This is declared as a custom test in Cargo.toml, one not using the
standard cargo harness. It must be enabled specifically when testing:

    cargo test --features=test-ui-replay

This solves the problem of not being able to run that test on macOS,
which requires GTK to be initialized and used only from the main thread.

The CI workflow is amended to test with the test-ui-replay feature
enabled, and to build the release binary without features enabled. All
current features are for debug and test purposes only.