Open our own libusb context rather than using the global one.
This is necessary to ensure an error dialog rather than a panic, in the
case where Packetry is run on a VirtualBox VM without USB forwarding.

In that scenario, libusb_init() returns an error, which is translated
into a panic by rusb, since the API does not allow for returning a
Result when GlobalContext is used.