Add hardware-in-the-loop test using Cynthion analyzer.
The test requires a Cynthion r0.6 or higher board, with the analyzer
gateware including a self-test capability.

The host must be connected to both the CONTROL and TARGET-C ports. The
TARGET-A port must be connected to the AUX port.

This test is optional and requires the 'test-cynthion' feature to be
enabled.