Pass error handler to capture thread, instead of errors through channel.
Currently, if errors occur during capture, the capture thread sends them
to the decoder thread, via the same channel used for packets.

This allows errors to be caught in the decoder thread and shown
immediately to the user, even when the the stop button has not yet been
pressed, and therefore the capture thread not yet joined.

This is unnecessarily complicated though, and can currently lead to some
errors being shown twice (where both sent to the decoder thread and
returned in the capture thread), or not being shown (where returned in
the capture thread, but not sent to the decoder thread).

Instead, let's just pass a handler function to LunaHandle::start(),
which will be called by the capture thread as it ends. If an error
occured, it is then signalled to the user immediately, without the
decoder thread needing to be involved.

When joining the capture thread, it is still necessary to check if the
thread panicked, but any normal error will already have been handled.

We can use the existing display_error() function, which already takes
care of accessing the UI from whichever thread it is called from.