Add packetry-cli wrapper program.
This is necessary on Windows, where we can't have the same binary serve
as both a GUI and a console program, due to the constraints of the PE
binary format.

Instead, we add a new 'packetry-cli' console binary which simply acts as
a wrapper around the main GUI binary.

The wrapper sets the PACKETRY_ATTACH_CONSOLE environment variable, which
tells the main process to attach to the wrapping parent's console. The
wrapper waits for the child to complete, and mirrors its exit code.