Fix compact_index test.
This test was copied and pasted from the test for index_stream, to the
extent where we also forgot to change the type being tested! This commit
corrects that.

Fortunately it turns out that compact_index does indeed work correctly,
although the test does trigger an underflow, which is not reachable in
normal usage of the application.

This commit fixes that underflow, which was caused by unnecessarily
subtracting 1 from a segment ID only to add it back again later.