Decode IN/OUT + DATA0/DATA1 based on endpoint type.
Now that we have that context in the transaction state, we can decode
IN | OUT followed by DATA0 | DATA1 to  one of three possibilities:

- Done if the endpoint is isochronous
- Continue if the endpoint is not isochronous
- Ambiguous if we don't know

In the Ambiguous case, we add the packet to the transaction, but don't
end the transaction or index any payload data. The transaction will be
ended when the next token starts a new transaction.