Ensure that we don't change n_items() without an items_changed signal.
In the ListModel interface, we must always report a n_items() that is
consistent with the history of `items_changed` signals we have emitted.

So keep a published row count separately from our internal counts, and
update it just before emitting an `items_changed` signal.

Fixes crashes in GTK where the view gets out of sync with the model.